// api/models/UserJWT.ts
/**
 * some api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class UserJWT {
    'user_jwt': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "user_jwt",
            "baseName": "user_jwt",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return UserJWT.attributeTypeMap;
    }

    public constructor() {
    }
}